+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] Q1 2023
+ - metrics_provider:
+   - reduced actix-web workers to 1
+   - refactored default address handling
+   - refactored drop behaviour
+
 ## [0.2.0] Q1 2023
  - replaced warp by actix-web
  - replaced BIND_ADDR and PORT by BIND_ADDR_AND_PORT